Moved jwt config constructor
